Task #31214 - Reworked LeaveRequestForm into 2 components depending on day/period.
